+++ StarkBank/StarkBank/User/User.cs
@@ -7,23 +7,18 @@ namespace StarkBank
     {
         public string Pem { get; }
         public string Environment { get; }
-        readonly private string Kind;
 
-        internal User(string environment, string id, string privateKey, string kind) : base(id)
+        internal User(string environment, string id, string privateKey) : base(id)
         {
             Pem = Checks.CheckPrivateKey(privateKey);
             Environment = Checks.CheckEnvironment(environment);
-            Kind = kind;
-        }
-
-        internal string AccessId()
-        {
-            return Kind + "/" + ID;
         }
 
         internal EllipticCurve.PrivateKey PrivateKey()
         {
             return EllipticCurve.PrivateKey.fromPem(Pem);
         }
+
+        internal abstract string AccessId();
     }
 }