+++ StarkBank/StarkBank/UtilityPayment/Log.cs
@@ -66,7 +66,7 @@ namespace StarkBank
             /// <br/>
             /// Parameters (optional):
             /// <list>
-            ///     <item>user [Project object]: Project object. Not necessary if StarkBank.User.Default was set before function call</item>
+            ///     <item>user [Organization/Project object]: Organization or Project object. Not necessary if StarkBank.Settings.User was set before function call</item>
             /// </list>
             /// <br/>
             /// Return: